refact : CreditLog 멤버 필드 생성 및 연동
CreditLog Entity에 Member 필드 생성
ChargeService의 addCredit 메서드와 테스트 데이터에 Member 추가
