+++ src/main/java/com/bitharmony/comma/credit/charge/dto/ChargeConfirmRequest.java
@@ -1,0 +1,12 @@
+package com.bitharmony.comma.credit.charge.dto;
+
+import lombok.Builder;
+
+@Builder
+public record ChargeConfirmRequest(
+        String paymentKey,
+        String orderId,
+        String amount
+
+        ) {
+}
