+++ src/main/java/com/bitharmony/comma/credit/charge/dto/ChargeConfirmResponse.java
@@ -1,0 +1,11 @@
+package com.bitharmony.comma.credit.charge.dto;
+
+import lombok.Builder;
+import org.json.simple.JSONObject;
+
+@Builder
+public record ChargeConfirmResponse(
+        JSONObject paymentStatement,
+        boolean isApproved
+) {
+}
