+++ src/main/java/com/bitharmony/comma/credit/charge/dto/ChargeCreateRequest.java
@@ -1,0 +1,9 @@
+package com.bitharmony.comma.credit.charge.dto;
+
+import lombok.Builder;
+
+@Builder
+public record ChargeCreateRequest(
+        long chargeAmount
+) {
+}
