+++ src/main/java/com/bitharmony/comma/credit/charge/dto/ChargeCreateResponse.java
@@ -1,0 +1,9 @@
+package com.bitharmony.comma.credit.charge.dto;
+
+import lombok.Builder;
+
+@Builder
+public record ChargeCreateResponse(
+        long chargeId
+) {
+}
