+++ src/main/java/com/bitharmony/comma/credit/charge/dto/ChargeGetListResponse.java
@@ -1,0 +1,11 @@
+package com.bitharmony.comma.credit.charge.dto;
+
+import lombok.Builder;
+
+import java.util.List;
+
+@Builder
+public record ChargeGetListResponse(
+        List<ChargeDto> chargeDtos
+) {
+}
