+++ src/main/java/com/bitharmony/comma/credit/charge/repository/ChargeRepository.java
@@ -1,0 +1,12 @@
+package com.bitharmony.comma.credit.charge.repository;
+
+import com.bitharmony.comma.credit.charge.entity.Charge;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface ChargeRepository extends JpaRepository<Charge, Long> {
+    List<Charge> findByChargerId(Long id);
+}
