+++ src/main/java/com/bitharmony/comma/credit/creditLog/repository/CreditLogRepository.java
@@ -1,0 +1,9 @@
+package com.bitharmony.comma.credit.creditLog.repository;
+
+import com.bitharmony.comma.credit.creditLog.entity.CreditLog;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CreditLogRepository extends JpaRepository<CreditLog, Long> {
+}
