+++ src/main/java/com/bitharmony/comma/credit/withdraw/repository/WithdrawRepository.java
@@ -1,0 +1,7 @@
+package com.bitharmony.comma.credit.withdraw.repository;
+
+import com.bitharmony.comma.credit.withdraw.entity.Withdraw;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface WithdrawRepository extends JpaRepository<Withdraw, Long>  {
+}
