--- src/main/java/com/bitharmony/comma/domain/credit/charge/dto/ChargeConfirmRequest.java
@@ -1,12 +1,0 @@
-package com.bitharmony.comma.domain.credit.charge.dto;
-
-import lombok.Builder;
-
-@Builder
-public record ChargeConfirmRequest(
-        String paymentKey,
-        String orderId,
-        String amount
-
-        ) {
-}
