--- src/main/java/com/bitharmony/comma/domain/credit/charge/dto/ChargeCreateRequest.java
@@ -1,10 +1,0 @@
-package com.bitharmony.comma.domain.credit.charge.dto;
-
-import com.bitharmony.comma.domain.credit.charge.entity.Charge;
-import lombok.Builder;
-
-@Builder
-public record ChargeCreateRequest(
-        long chargeAmount
-) {
-}
