--- src/main/java/com/bitharmony/comma/domain/credit/charge/dto/ChargeCreateResponse.java
@@ -1,12 +1,0 @@
-package com.bitharmony.comma.domain.credit.charge.dto;
-
-import com.bitharmony.comma.domain.credit.charge.entity.Charge;
-import lombok.Builder;
-
-import java.time.LocalDateTime;
-
-@Builder
-public record ChargeCreateResponse(
-        long chargeId
-) {
-}
