--- src/main/java/com/bitharmony/comma/domain/credit/charge/service/ChargeService.java
@@ -1,111 +1,0 @@
-package com.bitharmony.comma.domain.credit.charge.service;
-
-import com.bitharmony.comma.domain.credit.charge.dto.ChargeConfirmResponse;
-import com.bitharmony.comma.domain.credit.charge.entity.Charge;
-import com.bitharmony.comma.domain.credit.charge.repository.ChargeRepository;
-import com.bitharmony.comma.domain.credit.creditLog.entity.CreditLog;
-import com.bitharmony.comma.domain.credit.creditLog.service.CreditLogService;
-import com.bitharmony.comma.global.exception.ChargeAmountNotMatchException;
-import com.bitharmony.comma.global.exception.ChargeNotFoundException;
-import com.bitharmony.comma.global.exception.PaymentFailureException;
-import com.bitharmony.comma.member.entity.Member;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class ChargeService {
-
-    private final ChargeRepository chargeRepository;
-    private final CreditLogService creditLogService;
-    private final TossPaymentsService tossPaymentsService;
-
-
-    public Charge getChargeById(long id) {
-        Optional<Charge> charge = this.chargeRepository.findById(id);
-
-        if (charge.isEmpty()) {
-            throw new ChargeNotFoundException();
-        }
-
-        return charge.get();
-    }
-
-    public List<Charge> getChargeList() {
-        return this.chargeRepository.findAll();
-    }
-
-    public Charge createCharge(Member member, long chargeAmount) {
-        Charge charge = Charge.builder()
-                .charger(member)
-                .chargeAmount(chargeAmount)
-                .build();
-        this.chargeRepository.save(charge);
-
-        return charge;
-    }
-
-
-    // 주문서와 결제요청의 정보가 일치하는지 확인하는 메서드
-    public void checkValidity(String orderId, long amount) {
-        long id = Long.parseLong(orderId.split("__", 2)[1]);
-        Charge charge = chargeRepository.findById(id).orElse(null);
-
-        if (charge == null) {
-            throw new ChargeNotFoundException();
-        }
-
-        if (charge.getChargeAmount() != amount) {
-            throw new ChargeAmountNotMatchException();
-        }
-    }
-
-
-    // 결제 성공시 멤버의 크레딧을 증가시키고 크레딧 로그를 남기는 메서드 (추후 멤버 기능 연동시 수정)
-    public void addCredit(String orderId, long amount, String paymentKey) {
-        long id = Long.parseLong(orderId.split("__", 2)[1]);
-        Charge charge = chargeRepository.findById(id).orElse(null);
-
-        if (charge == null) {
-            throw new ChargeNotFoundException();
-        }
-
-        Charge _charge = charge.toBuilder()
-                .payDate(LocalDateTime.now())
-                .paymentKey(paymentKey)
-                .build();
-
-        chargeRepository.save(_charge);
-
-        creditLogService.addCreditLog(CreditLog.EventType.충전__토스페이먼츠, amount);
-    }
-
-
-    public ChargeConfirmResponse confirmPayment(String orderId, String amount, String paymentKey) {
-
-        // orderId와 amount 맞는지 체크하는 메서드
-        checkValidity(orderId, Long.parseLong(amount));
-
-
-        try {
-            ChargeConfirmResponse chargeConfirmResponse =
-                    tossPaymentsService.requestApprovalAndGetResponse(orderId, amount, paymentKey);
-
-            if(chargeConfirmResponse.isApproved()) {
-                addCredit(orderId, Long.parseLong(amount), paymentKey);
-            }
-
-            return chargeConfirmResponse;
-        } catch (Exception e) {
-            throw new PaymentFailureException();
-        }
-    }
-
-    public List<Charge> getChargeListByMemberId(Long id) {
-        return chargeRepository.findByChargerId(id);
-    }
-}
